created class Search
